+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## v0.0.58 (0/26/2020)
+## v0.0.59 (02/29/2020)
+
+- Introduced the `Export to Repository` command to the `Gists` tree, which allows you export any of your gists to new GitHub repositories
+- Added support for adding `import` statements to playgrounds, without needing to use the `.mjs` extension and/or setting the `scriptType` property in the `playground.json` file.
+
+## v0.0.58 (02/26/2020)
 
 - Fixed a bug where gists without descriptions weren't loading properly